task/replace-graph * Tweak a few things to please CI and also extract the graph transform from the visual component because the test infra doesn't understand es6 imports
