+++ blueocean-dashboard/package.json
@@ -62,7 +62,7 @@
     "@jenkins-cd/design-language": "0.0.162",
     "@jenkins-cd/js-extensions": "0.0.44",
     "@jenkins-cd/js-modules": "0.0.10",
-    "@jenkins-cd/pipeline-graph-widget": "^0.1.2",
+    "@jenkins-cd/pipeline-graph-widget": "0.1.2",
     "@jenkins-cd/preferences": "0.0.4",
     "babel-plugin-transform-decorators-legacy": "1.3.4",
     "bluebird": "3.5.0",
